Compute and visualize stereo optical flow.

+++ src/debug.rs
@@ -6,7 +6,7 @@ lazy_static! {
 
 #[derive(Default)]
 pub struct DebugData {
-  pub detections: Vec<Vector2i>,
+  pub detections: Vec<Vector2d>,
   pub detection_mask: Vec<bool>,
   pub flow: Vec<(Vector2d, Vector2d)>,
 }
